Add files via upload
Coronavirus disease (COVID-19) was an infectious disease caused by the SARS-CoV-2 virus. It spreads through the transfer of liquid particles from an infected person’s mouth or nose. Under prevention protocols, people wear masks, clean their hands, and maintain social distancing to prevent its spread. Because COVID was first discovered in China, racially motivated crimes against Asians have increased in the United States. My research question was ‘How does the spread of coronavirus affect the crime rates in California from 2016 to 2020?’. California, a state in the Western United States with diverse and nuanced demographic features, can be used as a sample to estimate the characteristics of the United States. Based on data from the State of California Department of Justice, this research paper seeks to analyze trends in crime progression, pre/post the spread of coronavirus and produce color-coded maps using geographic information systems (GIS) to showcase the results. This research will simultaneously explore the impact of other demographic features on the crime rates, such as ethnicity, education level, and income level. This question was worthy of investigation because it affects groups of people in California, enabling a greater understanding of the topic. Through the map plotting and trends analysis, it was discovered that the outbreak of COVID was correlated with a decrease in property crimes and an increase in the number of violent crimes. Most of the crimes may be bias-motivated crimes, and the drop in crime case numbers may be attributed to COVID restrictions.